+++ lib/extensions/context_extension.dart
@@ -5,13 +5,14 @@ import 'package:flutter/material.dart';
 extension UiBuildContextExtension on BuildContext {
   ThemeData get theme => Theme.of(this);
   FlutterView get view => View.of(this);
-  MediaQueryData get mq => MediaQuery.of(this);
 
-  double get height => mq.size.height;
-  double get width => mq.size.width;
+  Size get mqSize => MediaQuery.sizeOf(this);
+  double get height => mqSize.height;
+  double get width => mqSize.width;
 
-  bool get isDarkMode => mq.platformBrightness == Brightness.dark;
-  bool get isLightMode => mq.platformBrightness == Brightness.light;
+  Brightness get brightness => MediaQuery.platformBrightnessOf(this);
+  bool get isDarkMode => brightness == Brightness.dark;
+  bool get isLightMode => brightness == Brightness.light;
 
   bool get isPhone => width < 600;
   bool get isTablet => !isPhone && !isDesktop;