+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.1
+
+* Added minSize, maxSize, minDimensions, and maxDimensions to UiConstrained extension.
+* Fixed UiList ordering.
+
 ## 0.1.0
 
 * Added UiBordered through `bordered` library (exported).